Add files via upload
The Iris flowers classification project is a machine learning project that involves building and training a model to classify different species of iris flower based on their physical characteristics.The goal of the project is to accurately predict the species of an iris flower based on its sepal length, sepal width, petal length, and petal width.
The project involves several steps, including data preprocessing, model selection, and model training. The dataset used in the project is
http://archive.ics.uci.edu/ml/datasets/Iris
The project involves several steps, including data preprocessing, model selection, and model training. Next, a suitable machine learning model is selected and trained using the training set. There are several different models that can be used for this task, including decision trees, support vector machines, and neural networks. The performance of the model is evaluated using metrics such as accuracy, precision, recall, and F1 score.The model takes in the physical characteristics of an iris flower as input and outputs a predicted species.
Overall, the Iris Flowers Classification project is a great way to learn about machine learning and data science. It provides hands-on experience with data preprocessing, model selection, and model training, and allows you to build a real-world application that can be used to classify iris flowers.